refactor: move types to own files
