+++ custom_components/sourceful_zap/system_sensor_definitions.py
@@ -1,0 +1,110 @@
+"""System sensor definitions for Zap, mapping to the API structure."""
+
+from homeassistant.components.sensor import SensorDeviceClass, SensorStateClass
+
+SYSTEM_SENSOR_DEFINITIONS = {
+    "uptime": {
+        "name": "Uptime",
+        "unit": "s",
+        "device_class": SensorDeviceClass.DURATION,
+        "state_class": SensorStateClass.TOTAL_INCREASING,
+        "icon": "mdi:timer-outline",
+        "path": "uptime_seconds",
+    },
+    "temperature": {
+        "name": "Temperature",
+        "unit": "°C",
+        "device_class": SensorDeviceClass.TEMPERATURE,
+        "state_class": SensorStateClass.MEASUREMENT,
+        "icon": "mdi:thermometer",
+        "path": "temperature_celsius",
+    },
+    "memory_percent": {
+        "name": "Memory Usage",
+        "unit": "%",
+        "device_class": None,
+        "state_class": SensorStateClass.MEASUREMENT,
+        "icon": "mdi:memory",
+        "path": "memory_MB.percent_used",
+    },
+    "memory_used": {
+        "name": "Memory Used",
+        "unit": "MB",
+        "device_class": SensorDeviceClass.DATA_SIZE,
+        "state_class": SensorStateClass.MEASUREMENT,
+        "icon": "mdi:memory",
+        "path": "memory_MB.used",
+    },
+    "memory_free": {
+        "name": "Memory Free",
+        "unit": "MB",
+        "device_class": SensorDeviceClass.DATA_SIZE,
+        "state_class": SensorStateClass.MEASUREMENT,
+        "icon": "mdi:memory",
+        "path": "memory_MB.free",
+    },
+    "wifi_rssi": {
+        "name": "WiFi Signal Strength",
+        "unit": "dBm",
+        "device_class": SensorDeviceClass.SIGNAL_STRENGTH,
+        "state_class": SensorStateClass.MEASUREMENT,
+        "icon": "mdi:wifi-strength-2",
+        "path": "zap.network.rssi",
+    },
+    "firmware_version": {
+        "name": "Firmware Version",
+        "unit": None,
+        "device_class": None,
+        "state_class": None,
+        "icon": "mdi:chip",
+        "path": "zap.firmwareVersion",
+    },
+    "cpu_frequency": {
+        "name": "CPU Frequency",
+        "unit": "MHz",
+        "device_class": SensorDeviceClass.FREQUENCY,
+        "state_class": SensorStateClass.MEASUREMENT,
+        "icon": "mdi:cpu-32-bit",
+        "path": "zap.cpuFreqMHz",
+    },
+    "flash_size": {
+        "name": "Flash Size",
+        "unit": "MB",
+        "device_class": SensorDeviceClass.DATA_SIZE,
+        "state_class": SensorStateClass.MEASUREMENT,
+        "icon": "mdi:harddisk",
+        "path": "zap.flashSizeMB",
+    },
+    "wifi_status": {
+        "name": "WiFi Status",
+        "unit": None,
+        "device_class": None,
+        "state_class": None,
+        "icon": "mdi:wifi",
+        "path": "zap.network.wifiStatus",
+    },
+    "local_ip": {
+        "name": "Local IP",
+        "unit": None,
+        "device_class": None,
+        "state_class": None,
+        "icon": "mdi:ip-network",
+        "path": "zap.network.localIP",
+    },
+    "wifi_ssid": {
+        "name": "WiFi SSID",
+        "unit": None,
+        "device_class": None,
+        "state_class": None,
+        "icon": "mdi:router-wireless",
+        "path": "zap.network.ssid",
+    },
+    "device_id": {
+        "name": "Device ID",
+        "unit": None,
+        "device_class": None,
+        "state_class": None,
+        "icon": "mdi:identifier",
+        "path": "zap.deviceId",
+    },
+}
